Version_2

+++ src/Etudiant.java
@@ -5,11 +5,11 @@ import java.util.Map;
 
 public class Etudiant {
     String nom;
-    byte age;
-    byte noteMoyenne;
+    int age;
+    int noteMoyenne;
     List<Object> listeEtudiants;
 
-    public Etudiant(String nom, byte age, byte noteMoyenne) {
+    public Etudiant(String nom, int age, int noteMoyenne) {
         this.nom = nom;
         this.age = age;
         this.noteMoyenne = noteMoyenne;
@@ -19,7 +19,7 @@ public class Etudiant {
     public String getNom() {
         return this.nom;
     }
-    public byte getAge() {
+    public int getAge() {
         return this.age;
     }
